+++ vms/evm/predicate/predicate.go
@@ -11,7 +11,7 @@ import (
 	"github.com/ava-labs/libevm/core/types"
 )
 
-// Delimiter separates the actual predicate bytes from the padded zero bytes.
+// delimiter separates the actual predicate bytes from the padded zero bytes.
 //
 // Predicates are encoded in the Access List of transactions by using in the
 // access tuples. This means that the length must be a multiple of
@@ -19,17 +19,35 @@ import (
 //
 // Even if the original predicate bytes is a multiple of [common.HashLength],
 // the delimiter must be appended to support decoding.
-const Delimiter = 0xff
+const delimiter = 0xff
 
 var (
-	ErrMissingDelimiter  = errors.New("no delimiter found")
-	ErrExcessPadding     = errors.New("predicate included excess padding")
-	ErrWrongEndDelimiter = errors.New("wrong delimiter")
+	errMissingDelimiter = errors.New("no delimiter found")
+	errExcessPadding    = errors.New("predicate included excess padding")
+	errWrongDelimiter   = errors.New("wrong delimiter")
 )
 
 // Predicate is a message padded with the delimiter and zeros and chunked into
 // 32-byte chunks.
 type Predicate []common.Hash
+
+// New constructs a predicate from raw predicate bytes.
+//
+// It chunks the predicate by appending [predicate.Delimiter] and zero-padding
+// to a multiple of 32 bytes.
+func New(b []byte) Predicate {
+	numUnpaddedChunks := len(b) / common.HashLength
+	chunks := make([]common.Hash, numUnpaddedChunks+1)
+	// Copy over chunks that don't require padding.
+	for i := range chunks[:numUnpaddedChunks] {
+		chunks[i] = common.Hash(b[common.HashLength*i:])
+	}
+
+	// Add the delimiter and required padding to the last chunk.
+	copy(chunks[numUnpaddedChunks][:], b[common.HashLength*numUnpaddedChunks:])
+	chunks[numUnpaddedChunks][len(b)%common.HashLength] = delimiter
+	return chunks
+}
 
 // Bytes converts the chunked predicate into the original message.
 //
@@ -41,17 +59,17 @@ func (p Predicate) Bytes() ([]byte, error) {
 	}
 	trimmed := common.TrimRightZeroes(padded)
 	if len(trimmed) == 0 {
-		return nil, fmt.Errorf("%w: length (%d)", ErrMissingDelimiter, len(p))
+		return nil, fmt.Errorf("%w: length (%d)", errMissingDelimiter, len(p))
 	}
 
 	expectedLen := (len(trimmed) + common.HashLength - 1) / common.HashLength
 	if expectedLen != len(p) {
-		return nil, fmt.Errorf("%w: got length (%d), expected length (%d)", ErrExcessPadding, len(p), expectedLen)
+		return nil, fmt.Errorf("%w: got length (%d), expected length (%d)", errExcessPadding, len(p), expectedLen)
 	}
 
 	delimiterIndex := len(trimmed) - 1
-	if trimmed[delimiterIndex] != Delimiter {
-		return nil, ErrWrongEndDelimiter
+	if trimmed[delimiterIndex] != delimiter {
+		return nil, errWrongDelimiter
 	}
 
 	return trimmed[:delimiterIndex], nil
